(GH-10) Author tool config simple script
In this commit a tool configuration has been added for the purpose of
testing `prm exec` functionality. When the tool is executed, it will return the contents of
the code directory it has been provided to execute upon.

In addition, adding the `-f` flag to the `toolArgs` will cause the tool
to exit with an exit code of 1, for testing purposes.